requirements.txt
# JARVIS AI Assistant - Python Dependencies

# Core ML and AI
torch>=2.0.0
numpy>=1.22.0
nemo_toolkit[all]>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.0.0

# Audio Processing
pyaudio>=0.2.11
soundfile>=0.12.0
librosa>=0.10.0

# Web Framework
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# HTTP and Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# AI and Language Processing
openai>=1.0.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Data Processing
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.1.0

# System and Utilities
psutil>=5.9.0
python-dotenv>=1.0.0
pyjokes>=0.6.0
wikipedia-api>=0.8.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0

# Additional utilities
tqdm>=4.64.0
packaging>=21.0
setuptools>=65.0.0
wheel>=0.37.0
